- integration tests moved to tests/ folder - pretty_assertions added
