+++ Cargo.toml
@@ -18,6 +18,7 @@ proptest = "1.1.0"
 criterion = "0.4"
 const-random = "0.1.15"
 iai = "0.1"
+pretty_assertions = "1.3"
 
 [lib]
 crate-type = ["cdylib", "staticlib", "lib"]
