+++ bench/criterion_bench.rs
@@ -4,7 +4,8 @@ use lambdaworks_kzg::{
 };
 
 fn bench_read_srs(c: &mut Criterion) {
-    const TRUSTED_SETUP_FILE: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/test/trusted_setup.txt");
+    const TRUSTED_SETUP_FILE: &str =
+        concat!(env!("CARGO_MANIFEST_DIR"), "/tests/trusted_setup.txt");
 
     let s = load_trusted_setup_file(TRUSTED_SETUP_FILE.lines()).unwrap();
     c.bench_function("kzgsettings_to_ref_string", |b| {
