+++ tests/trusted_setup_4.txt
